Began adding support for new market data
Added the prediction of new data from the yfinance library. Accuracy was close to 99.7%. Worried that the model is simply using the last price as the prediction for the next one. Not entirely true, as seen in the final graph. However, still an issue as the model seems to exhibit this behavior in some areas.